Adding version numbers.
Since old binaries continue to be copied around, it could help to
be able to identify different versions of the software.

The version number is be shown at the top of the "-h" usage message
and also at the bottom of the "-vvh" output.

At this time, the version is "1.0.dev" which is meant to imply
"a development version some time after 1.0".

Soon I intend to have a version "1.1" to be followed by "1.1.dev".